// frontend/src/app/shared/api/generated/model/updateTaskRequest.ts
/**
 * template API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateTaskRequest { 
    id?: number | null;
    name: string;
    description: string;
    checked?: boolean;
}